Crate progress

progress is meant to be a set of useful tools for showing program running progress (as its name) and steps.

Installation

Add the following lines to your Cargo.toml dependencies section, if you use Cargo:

[dependencies]
progress = "0.1.0"

Usage

Please check documentations for each structs. Life is easy here :)

Who create this

Ying-Ruei Liang (KK)

Contribution

I can't believe you would say that, but if you have any great idea or any bug report, don't be hesitate! It would be more wonderful if someone wants to write some code for this project!

TODO list

BarBuilder, so we can do some customization, e.g. change the symbols used

Add more type of indicators, e.g. spinning symbol or nayn cat :3

Color/styled text support (print!("{:<50}") will count unprintable text as well, I have to solve it first)

Make output format customizable, despite I have no idea how to achieve this for now.

License

MIT

Structs

Bar
Struct that used for presenting progress bar with plain texts.

BarBuilder
A builder that used for creating customize progress bar.

SpinningCircle
Struct that used for presenting progress with plain texts.

Text
Struct that used for presenting progress with plain texts.