Struct progress::SpinningCircle

pub struct SpinningCircle {
    // some fields omitted
}

Struct that used for presenting progress with plain texts.

It looks like:

* Doing something
/ Doing another thing

Examples

use std::thread;

extern crate progress;

fn main() {
    let mut spinningCircle = progress::SpinningCircle::new();

    spinningCircle.set_job_title("Writing boring and stupid homeworks");
    for _ in 0..50 {
        thread::sleep_ms(50);
        spinningCircle.tick();
    }
    spinningCircle.jobs_done();

    spinningCircle.set_job_title("Previewing boring and stupid subjects");
    for _ in 0..50 {
        thread::sleep_ms(50);
        spinningCircle.tick();
    }
    spinningCircle.jobs_done();

    spinningCircle.set_job_title("Learning and creating interesting programs");
    for _ in 0..50 {
        thread::sleep_ms(50);
        spinningCircle.tick();
    }
    spinningCircle.jobs_done();
}

Methods

impl SpinningCircle

fn new() -> SpinningCircle

Create a new progress spinning circle.

fn set_job_title(&mut self, new_title: &str)

Set text shown in progress spinning circle.

fn tick(&mut self)

Tell spinning circle to spin a bit.

fn jobs_done(&self)

Print finished symbol at the position spinning circle symbol used to be. And print "\n" to jump to next line.

e.g. * Collection kitties